pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fusectl"
version = "0.1.0"
description = "Gerenciador Linux para Custom Firmware Nintendo Switch"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
    "Topic :: Utilities",
]
dependencies = [
    "pyusb>=1.2.0",
    "textual>=0.40.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
fusectl = "fusectl.__main__:main"

[tool.setuptools.packages.find]
include = ["fusectl*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
